Revert "refactor: prettify import paths"
This reverts commit 23471dd7d50792321b0487affcab943eeafade08.
