+++ packages/digitraffic-mqtt/src/index.ts
@@ -1,3 +1,3 @@
-export * from '~/handlers/subscribe_to_station'
-export * from '~/handlers/subscribe_to_trains'
-export * from '~/handlers/subscribe_to_train_locations'
+export * from './handlers/subscribe_to_station.js'
+export * from './handlers/subscribe_to_trains.js'
+export * from './handlers/subscribe_to_train_locations.js'
