+++ packages/digitraffic-mqtt/tests/handlers/subscribe_to_station.test.ts
@@ -1,16 +1,18 @@
 import type { Train } from '@junat/digitraffic/types'
 
-import type { StationMqttClient } from '~/handlers/subscribe_to_station'
+import type { StationMqttClient } from '../../src/types/station_mqtt_client'
 
 import { describe, it, expect, beforeAll, afterAll } from 'vitest'
 
-import { subscribeToStation } from '~/handlers/subscribe_to_station'
+import { subscribeToStation } from '../../src/handlers/subscribe_to_station'
 
 describe('subscribe to station', () => {
   let client: StationMqttClient
 
   beforeAll(async () => {
-    client = await subscribeToStation('HKI')
+    client = await subscribeToStation({
+      stationShortCode: 'HKI'
+    })
   })
 
   it('connects', async () => {
