+++ packages/digitraffic-mqtt/vite.config.ts
@@ -4,9 +4,6 @@ export default defineConfig({
   test: {
     globalSetup: ['./src/mqtt_server_setup.ts'],
     setupFiles: ['./src/setup.ts'],
-    alias: {
-      "~": "src"
-    },
     coverage: {
       all: true,
       include: ['src'],
